Check classification result in "make test"
Make tests/test.sh check the result. As the test itself is written in
shell, just use shell ways to check it - instead of actually parsing the
XML.

The classification check also assumes that the classification results
are sorted by descending probabilities, i.e. that the result with
highest probability is listed first.

Fixes #3.